#23 Fix order of alert and correct answer
When the end user of the wio terminal publishes their answer, the correct/incorrect answer will be displayed directly on the display now, instead of waiting for the teacher to press Ok on the alert. This commit also changes how the checkAnswer parameter works. We now compare the animal.name property instead of the animal object.